Support requests with different types of payload
Until now, there was no possibility in the code to support different types
of requests, such as multipart form data requests, due to how the code
assumed that every request is involving a JSON payload.
This patch changes this behaviour, by allowing different types of requests
to be passed to retrieving functions, such as _get / _post / _put etc.
A new base class, BaseRequest, was introduced, which knows how to prepare itself
in order to be sendable to the remote API. There are multiple implementations
of it, for instance for JSON and multipart form data payloads. The API wrapper
methods are now accepting an instance of this request class, preparing it
before sending the content to the HTTP API.